render layer modifications

+++ src/MeshConvert.cpp
@@ -124,8 +124,8 @@ void saveMeshToFile(FILE *f)
         .magicValue = 0x12345678,
         .meshCount = (uint32_t)m_meshes.size(),
         .dataBlockStartOffset = (uint32_t)(sizeof(MeshFileHeader) + m_meshes.size() * sizeof(Mesh)),
-        .indexDataSize = m_indexData.size() * sizeof(uint32_t),
-        .vertexDataSize = m_vertexData.size() * sizeof(float)
+        .indexDataSize = static_cast<uint32_t>(m_indexData.size() * sizeof(uint32_t)),
+        .vertexDataSize = static_cast<uint32_t>(m_vertexData.size() * sizeof(float))
     };
 
     fwrite(&header, 1, sizeof(header), f);
